move func tests to own file and test verbose flag
